refactor: unify core API and reproduce paper evaluation protocols
- transforms/utils: deterministic Voxelize(reduce='first') + 33 regenerated output snapshots,
  RelabelBoxes ignore_mapping, DataKeys velocity/num_points/attribute, load_state_dict strict opt-out
- datasets: train:bool -> split:str (ScanObjectNN partition=), atomic caches, new ModelNet40Hdf5,
  nuScenes GT velocity/num_lidar_pts/attribute
- layers/losses: layer_container/reshape removed, block fixes, shared loss utils, chamfer distance
- models: configure_head/reset_classifier everywhere, {arch}.{dataset}.{author} naming, registry
  suggestions, point_transformer TransitionDown rebuilt, spvcnn fusion layer, velocity-aware decodes
- evaluation: torch-native exact BEV IoU, KITTI R11/R40 AP with per-class ignore regions, official
  nuScenes distance-mAP/TP-errors/NDS, indoor instance mask-AP; benchmarks follow the original
  authors' protocols, measured numbers labeled pre-fix pending re-measure
- examples: per-batch OneCycleLR stepping, seeding + tf32-off determinism, registry naming

+++ src/torch_pointcloud/utils/neighbors.py
@@ -20,7 +20,7 @@ def gaussian_kernel_density(x: Tensor, batch: Tensor, bandwidth: float) -> Tenso
         from torch_pointcloud.utils.neighbors import gaussian_kernel_density
 
         pos = torch.randn(100, 3)
-        batch = torch.zeros(100)
+        batch = torch.zeros(100, dtype=torch.long)
         bandwidth = 0.1
 
         density = gaussian_kernel_density(pos, batch, bandwidth)
@@ -31,8 +31,6 @@ def gaussian_kernel_density(x: Tensor, batch: Tensor, bandwidth: float) -> Tenso
 
     # TODO: Maybe avoid the expansive 'to_dense_batch' operation by using a better suited approach for packed data,
     # TODO: like a custom CUDA kernel to compute directly the KDE?
-    # NOTE: This function is approximately 20% slower than the original one (from PointConv paper)
-    # https://github.com/DylanWusee/pointconv_pytorch/blob/master/utils/pointconv_util.py
     x_dense, mask = to_dense_batch(x, batch)
 
     dist = -2 * torch.matmul(x_dense, x_dense.transpose(1, 2))
